requirements.txt
arrow==1.3.0
asgiref==3.8.1
bizerror==0.4.11
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
diff-match-patch==20230430
Django==5.0.6
django-click==2.4.0
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
django-environ==0.11.2
django-import-export==4.0.7
django-redis==5.4.0
django-rest-passwordreset==1.4.1
django-rq==2.10.2
django-safe-settings==0.1.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
fastutils==0.43.4
idna==3.10
jsonformatter==0.3.2
msgpack==1.1.0
pillow==10.3.0
psutil==6.0.0
psycopg2==2.9.9
pyasn1==0.6.1
pycryptodome==3.20.0
PyJWT==2.9.0
pypinyin==0.53.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.4
requests==2.32.3
rq==1.16.2
rq-win @ git+https://github.com/michaelbrooks/rq-win.git@c6e65e38f8a6ca99a0c8e0dd6fe35f9b92837cca
rsa==4.9
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
times==0.7
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
urllib3==2.2.3
zenutils==0.5.5
